refactor: extract archive ZIP reader into archive-extract.ts
Move the dual-format ZIP reader (extractFolderArchive, detectZipFormat,
extractNewFormatZip, extractOldFormatZip, extractImageFromZip, blobToDataUrl)
plus the reader-only helpers (getJSZip, LIMITS, IMAGE_EXTENSIONS, isImageFile,
ZipFormat) into a new utils/archive-extract.ts. archive.ts retains the writer
and re-exports getJSZip and extractFolderArchive so consumers importing from
$lib/utils/archive are unchanged. archive-extract.ts has no import from
archive.ts, so there is no import cycle.
